requirements.txt
requests
python-telegram-bot
loguru

